id from Update* and Telebot* changed to int64

+++ src/telebot/private/types.nim
@@ -37,7 +37,7 @@ type
     inlineQueryCallbacks*: seq[InlineQueryCallback]
     serverUrl*: string
     proxy*: Proxy
-    id*: int
+    id*: int64
     username*: string
 
   Command* = object
@@ -916,7 +916,7 @@ type
     retryAfter*: int
 
   Update* = ref object of TelegramObject
-    updateId*: int
+    updateId*: int64
     message*: Message
     editedMessage*: Message
     channelPost*: Message
